Highlighting active tab works
