+++ beerweb/ui/package.json
@@ -34,7 +34,7 @@
     "material-ui": "^0.15.0",
     "react-addons-css-transition-group": "^15.1.0",
     "react-hot-loader": "^1.3.0",
-    "react-router": "^2.0.0",
+    "react-router": "^2.4.1",
     "react-tap-event-plugin": "^1.0.0",
     "style-loader": "^0.13.0",
     "webpack": "^1.12.13",
